Add test & Fix some bugs

+++ app/UseCases/UnfollowUseCase.php
@@ -22,6 +22,7 @@ class UnfollowUseCase
         $targetId = user($targetUserId)->id;
 
         if (Follow::where('user_id', $id)->where('target_user_id', $targetId)->doesntExist()) return false;
+        if ($id === $targetId) return false;
 
         Follow::where('user_id', user()->id)->where('target_user_id', user($targetUserId)->id)->delete();
